cw: update, document and cache last_id
Rework the python web scraper script to be more resilient, taking into
account failures.

Also set the upper bounds to a high number -- and bail out on three
consecutive retries, if the crossword cannot be found.  This should last
for well into the year, until it needs to be changed again.

Cache the last fetched id and use that to offset where to start from
when looking for the next crossword.